Added integration test for homepage.
